refactor(provisioner): promote provisioner to internal/provisioner
Move-only: a leaf with zero internal imports and a single consumer, cmd/gpu.go.
Nothing about it was ever a shared service.
